add auth & update routes

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@description Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@description Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request body is:", req.body);
     const { name, email, phone } = req.body;
@@ -22,14 +22,15 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
     res.status(201).json(contact);
 });
 
 //@description Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -41,12 +42,16 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@description Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user contacts");
     }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, { new: true });
     res.status(201).json(updatedContact);
@@ -54,12 +59,16 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@description Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user contacts");
     }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(204).json({ message: `Delete contact for ${req.params.id}` });
